Frontend/monitoring-app/src/app/sidenav/sidenav.component.spec.ts
import { ComponentFixture} from '@angular/core/testing';

import { TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';
import { MatButtonModule } from '@angular/material/button';
import { MatIconModule } from '@angular/material/icon';
import { MatListModule } from '@angular/material/list';
import { MatSidenavModule } from '@angular/material/sidenav';
import { MatToolbarModule } from '@angular/material/toolbar';
import { CommonModule } from '@angular/common';
import { SidenavComponent } from './sidenav.component';


describe('SidenavComponent', () => {
  let component: SidenavComponent;
  let fixture: ComponentFixture<SidenavComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [SidenavComponent,
        RouterTestingModule,  
        MatButtonModule,
        MatIconModule,
        MatListModule,
        MatSidenavModule,
        MatToolbarModule,
        CommonModule]
    })
    .compileComponents();

    fixture = TestBed.createComponent(SidenavComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

  it('should create', () => {
    const fixture = TestBed.createComponent(SidenavComponent);
    const component = fixture.componentInstance;
    expect(component).toBeTruthy();
  });

  it('should set extendedNav to true when input is true', () => {
    const fixture = TestBed.createComponent(SidenavComponent);
    const component = fixture.componentInstance;
    component.extendedNav = true;
    fixture.detectChanges();
    expect(component.extendedNav).toBeTrue();
  })

  it('should toggle extendedNav when extendNav is called', () => {
    const fixture = TestBed.createComponent(SidenavComponent);
    const component = fixture.componentInstance;
  
    component.extendedNav = false;
    component.extendNav();
    expect(component.extendedNav).toBeTrue();
  
    component.extendNav();
    expect(component.extendedNav).toBeFalse();
  });
  

});
